Move MvvmCommand class to separate file
