+++ Miru/Miru.csproj
@@ -156,6 +156,7 @@
     <Compile Include="Views\Converters\InverseBooleanConverter.cs" />
     <Compile Include="Views\Extensions\RunVisibility.cs" />
     <Compile Include="ViewModels\Interfaces\ISimpleContentDialog.cs" />
+    <Compile Include="Views\MvvmCommand.cs" />
     <Compile Include="Views\ShellView.xaml.cs">
       <DependentUpon>ShellView.xaml</DependentUpon>
     </Compile>
